fix: change Stat()'s Links() retval from a uint64 to a uint

+++ lstat_test.go
@@ -133,7 +133,7 @@ func TestLstatLinks(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	var want uint64 = 1
+	var want uint = 1
 	if got := fi.Links(); got != want {
 		t.Fatalf("Links(): got %v, want %v", got, want)
 	}
